requirements.txt
Flask>=2.3.0
Flask-Login>=0.6.4
Flask-SQLAlchemy>=3.1.1
openai>=1.55.1
python-dotenv>=0.19.0
Werkzeug>=2.0.1
Pillow>=8.3.1
requests>=2.26.0
mysql-connector-python>=9.0.0